+++ Source/MACLib/UnBitArrayBase.h
@@ -31,7 +31,7 @@ public:
     virtual int FillAndResetBitArray(int64 nFileLocation = -1, int64 nNewBitIndex = 0);
         
     virtual void GenerateArray(int * pOutputArray, int nElements, intn nBytesRequired) {}
-    virtual uint64 DecodeValue(DECODE_VALUE_METHOD DecodeMethod, int nParam1 = 0, int nParam2 = 0) { return 0; }
+    virtual uint32 DecodeValue(DECODE_VALUE_METHOD DecodeMethod, int nParam1 = 0, int nParam2 = 0) { return 0; }
     
     virtual void AdvanceToByteBoundary();
     virtual bool EnsureBitsAvailable(uint32 nBits, bool bThrowExceptionOnFailure);
